+++ src/hooks/mutations/useCreateStudy.ts
@@ -1,5 +1,5 @@
 import studyApi from '@apis/study/studyApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 
@@ -14,12 +14,6 @@ export const useCreateStudy = () => {
       showToast({
         type: 'success',
         message: '글을 작성했어요.',
-      });
-    },
-    onError: () => {
-      showToast({
-        type: 'fail',
-        message: '오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
       });
     },
   });
