+++ src/hooks/mutations/useDeleteApplication.ts
@@ -1,5 +1,5 @@
 import applicationApi from '@apis/application/applicationApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { ApplicationParamType } from '@hooks/mutations/usePostApplication';
