+++ src/hooks/mutations/useDeleteMember.ts
@@ -1,5 +1,5 @@
 import memberApi from '@apis/member/memberApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { TOAST_MESSAGE } from '@constants/index';
