+++ src/hooks/mutations/useDeleteStudy.ts
@@ -1,5 +1,5 @@
 import studyApi from '@apis/study/studyApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 export const useDeleteStudy = () => {
