+++ src/hooks/mutations/useDeleteUser.ts
@@ -1,6 +1,6 @@
 import userApi from '@apis/user/userApi';
 import useUser from '@hooks/useUser';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation } from '@tanstack/react-query';
 
 export const useDeleteUser = () => {
@@ -14,12 +14,6 @@ export const useDeleteUser = () => {
       showToast({
         type: 'success',
         message: '회원 탈퇴가 성공적으로 처리되었습니다.',
-      });
-    },
-    onError: () => {
-      showToast({
-        type: 'fail',
-        message: '오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
       });
     },
   });
