+++ src/hooks/mutations/useEditStudy.ts
@@ -1,5 +1,5 @@
 import studyApi from '@apis/study/studyApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/router';
 
@@ -27,10 +27,11 @@ export const useEditStudy = () => {
       });
     },
     onError: () => {
-      showToast({
-        type: 'fail',
-        message: '오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
-      });
+      // showToast({
+      //   type: 'fail',
+      //   message: '오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
+      // });
+      router.back();
     },
   });
 };
