+++ src/hooks/mutations/usePutUser.ts
@@ -1,5 +1,5 @@
 import userApi from '@apis/user/userApi';
-import useToast from '@hooks/useToast';
+import { useToast } from '@providers/ToastProvider';
 import { userState } from '@recoil/auth';
 import { useMutation } from '@tanstack/react-query';
 import { useSetRecoilState } from 'recoil';
@@ -17,12 +17,6 @@ export const usePutUser = () => {
       showToast({
         type: 'success',
         message: '프로필을 저장했어요.',
-      });
-    },
-    onError: () => {
-      showToast({
-        type: 'fail',
-        message: '오류가 발생했습니다. 잠시 후 다시 시도해주세요.',
       });
     },
   });
